Updated Stage and created an EndlessLevelHandler.cs
The EndlessLevelHandler, contains a GetRandomSectionFromPool, which basically selects a random section from the sectionsPool that is not active (using activeInHierarchy). An UpdateSectionPositions which keeps a track of the last z position used and places the random section at this position. An UpdateLessOften which is called as a co-routine every 100ms. And a Start routine which fills the sectionsPool and then places a random section in the first position.