Added the AI Car to the game.
Added the AICar Variant.prefab as this uses near enough all the same mechanics as the Car.prefab.
Added a LayerMask and an array of colliders to the AICarSpawner, used these to check that we are not spawning new AICars ontop of other AICars.
Added an AIHandler and checked that this script is not running on a car tagged with Player, also done this on CarHandler but the other way around. Used RaycastHits to check that the current car is not being spawned ontop of another car. Set speed, a lot lower than the player, and lanes to make it so that you have to dodge the AICars.