Divided the Taxi into quarters using Blender and created AICarTaxi Variant.prefab
Used blender to divide the Taxi into quarters and added it as 'AICarTaxi Variant.prefab'.
Altered the AICarSpawner.cs to call SpawnNewCars twice once creating an AICar that is 60 ahead, and the other 80 ahead. Then allow the UpdateLessOften co-routine to create all other cars 100 ahead. Also altered the y position of the spawn dependent on the value recorded in the prefab (therefore we don't get floating or sunken cars).
To rectify cars driving in the middle of the road, reduced the second parameter to Mathf.Abs(difference... from 0.1f to 0.02f. And gave the SetMaxSpeed froat params of 2.0f & 4.8f.
Commented out a Debug.Log in CarHandler.cs & ExplodeHandler.cs.
Commented out the Way the project responds to KeyCode.R in InputHandler.cs.