Made the following modifications:
AICarSpawner.cs - It is not appropriate to spawn AI Cars if the Car isExploded.
FuelGauge.cs - commented out the pressing of the space key to get more fuel.
- didn't like the method for returning the fuelFillImage.color.
CarHandler.cs - made isExploded public and initialised the leaderboard.
GameOverScreen.cs - Tried to use highScoreManager not yet successfully.
Added a HighScoreManager.cs.